% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_inference.R
\name{auc_inference}
\alias{auc_inference}
\title{Estimate AUC and 95 CI}
\usage{
auc_inference(true, pred, alpha = 0.05)
}
\arguments{
\item{true}{Vector of true values, 0 or 1}

\item{pred}{Vector of predicted values used to sort the data}

\item{alpha}{Type I error rate of the statistical test}
}
\description{
Estimate AUC and 95 CI
}
