% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_table.R
\name{auc_table}
\alias{auc_table}
\title{Generate a table of AUCs by learner}
\usage{
auc_table(x, ...)
}
\arguments{
\item{x}{SuperLearner or CV.SuperLearner object}

\item{...}{Remaining arguments passed through to the methods.}
}
\description{
Supports both CV.SuperLearner and SuperLearner objects.
}
