% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_score.R
\name{brier_score}
\alias{brier_score}
\title{Calculate Brier score from cross-validation results}
\usage{
brier_score(preds, actual, test_folds)
}
\arguments{
\item{preds}{Predicted values or score}

\item{actual}{Known label value}

\item{test_folds}{Length n vector assigning each vector to a given test fold}
}
\value{
Dataframe of results, including Brier score and standard deviation
}
\description{
Calculate Brier score from cross-validation results
}
