% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_table.CV.SuperLearner.R
\name{brier_table.CV.SuperLearner}
\alias{brier_table.CV.SuperLearner}
\title{Table of cross-validated Brier scores from CV.SuperLearner result}
\usage{
\method{brier_table}{CV.SuperLearner}(
  x,
  y = x$Y,
  sort = TRUE,
  null_hypothesis = NULL,
  two_tailed = FALSE,
  lower.tail = TRUE,
  ...
)
}
\arguments{
\item{x}{CV.SuperLearner object}

\item{y}{Outcome vector, if not already added to CV.SL object.}

\item{sort}{Sort table by order of score.}

\item{null_hypothesis}{Not implemented yet}

\item{two_tailed}{Not implemented yet}

\item{lower.tail}{Not implemented yet}

\item{...}{Any additional unused arguments, due to the prauc_table generic.}
}
\value{
Dataframe table with Brier score and std dev.
}
\description{
Calculates cross-validated Brier score for each learner in the CV.SuperLearner.
Also calculates standard-error, confidence interval and p-value.
}
\examples{
library(SuperLearner)
library(ck37r)

data(Boston, package = "MASS")

set.seed(1)
y = as.numeric(Boston$medv > 23)
cvsl = CV.SuperLearner(Y = y,
                       X = subset(Boston, select = -medv),
                       family = binomial(),
                       cvControl = list(V = 2, stratifyCV = TRUE),
                       SL.library = c("SL.mean", "SL.glm"))
brier_table(cvsl, y = y)



}
