% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_table.R
\name{brier_table}
\alias{brier_table}
\title{Generate a table of Brier scores by learner}
\usage{
brier_table(x, ...)
}
\arguments{
\item{x}{CV.SuperLearner object}

\item{...}{Remaining arguments passed through to the methods.}
}
\description{
Supports CV.SuperLearner objects currently.
}
