% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoricals_to_factors.R
\name{categoricals_to_factors}
\alias{categoricals_to_factors}
\title{Convert categorical numerics to factors}
\usage{
categoricals_to_factors(data, categoricals, verbose = F)
}
\arguments{
\item{data}{Dataframe}

\item{categoricals}{Vector of column names that are categoricals.}

\item{verbose}{If T display extra output during execution.}
}
\value{
Updated dataframe.
}
\description{
Convert categorical numerics to factors
}
