% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{Create a dataset}
\usage{
create_dataset(
  data,
  varsX = "",
  varY = "",
  config_code = NULL,
  rename_vars = c(),
  remove_vars = c(),
  subset = "all",
  ylevel = NULL,
  primary_key = NULL,
  verbose = F
)
}
\arguments{
\item{data}{Dataframe with X and Y variables.}

\item{varsX}{Vector of variables to use as predictors.}

\item{varY}{Variable name to use as the outcome.}

\item{config_code}{R code to create varsX and varY, which define the
variables to use.}

\item{rename_vars}{named vector, where names are the original names and
values are the new names.}

\item{remove_vars}{vector of strings for variables to remove from varsX}

\item{subset}{Code to further subset the data.}

\item{ylevel}{Target level of y if y is a factor variable.}

\item{primary_key}{Id variable, if any.}

\item{verbose}{If TRUE display extra output.}
}
\description{
To be transferred into a separate package for Alan.
}
