% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvsl_auc.R
\name{cvsl_auc}
\alias{cvsl_auc}
\title{Calculate cross-validated AUC from CV.SuperLearner result}
\usage{
cvsl_auc(cvsl)
}
\arguments{
\item{cvsl}{CV.SuperLearner object}
}
\value{
List with cvAUC and ci elements.
}
\description{
Also calculates confidence interval.
Based on initial code by Alan Hubbard.
}
\examples{

library(SuperLearner)
library(ck37r)
data(Boston, package = "MASS")

set.seed(1)
cvsl = CV.SuperLearner(Boston$chas, subset(Boston, select = -chas), family = binomial(),
                      cvControl = list(V = 2, stratifyCV = TRUE),
                      SL.library = c("SL.mean", "SL.glmnet"))
cvsl_auc(cvsl)

}
\references{
LeDell, E., Petersen, M., & van der Laan, M. (2015). Computationally
efficient confidence intervals for cross-validated area under the ROC curve
estimates. Electronic journal of statistics, 9(1), 1583.
}
\seealso{
sl_auc cvAUC::ci.cvAUC
}
