% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvsl_auc_table.R
\name{cvsl_auc_table}
\alias{cvsl_auc_table}
\title{Deprecated wrapper to auc_table.CV.SuperLearner()}
\usage{
cvsl_auc_table(cvsl, ...)
}
\arguments{
\item{cvsl}{CV.SuperLearner object}
}
\value{
Dataframe table with auc, se, ci, and p-value.
}
\description{
Deprecated, please use
}
