% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvsl_plot_roc.R
\name{cvsl_plot_roc}
\alias{cvsl_plot_roc}
\title{Plot a ROC curve from cross-validated AUC from CV.SuperLearner}
\usage{
cvsl_plot_roc(cvsl, Y = cvsl$Y,
  title = "CV-SuperLearner Cross-validated ROC", digits = 4)
}
\arguments{
\item{cvsl}{tbd}

\item{Y}{tbd}

\item{title}{tbd}

\item{digits}{tbd}
}
\description{
Thank you to Alan Hubbard for the initial code.
}
\references{
Add Erin LeDell paper/chapter.
}
