% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvsl_weights.R
\name{cvsl_weights}
\alias{cvsl_weights}
\title{Create a table of meta-weights from a CV.SuperLearner}
\usage{
cvsl_weights(cvsl, sort = T, nonzero = F, clean_names = T, rank = T,
  digits = 5)
}
\arguments{
\item{cvsl}{CV.SuperLearner result object}

\item{sort}{If TRUE, sort rows (learners) in descending order by mean weight.}

\item{nonzero}{Restrict to learners with a non-zero mean weight.}

\item{clean_names}{Remove "SL." from the front and "_All" from the end of
learner names.}

\item{rank}{Adding the learner rank to the table.}

\item{digits}{Number of digits to round the results. Set to NULL to disable.}
}
\value{
Table in data frame form with each learner's mean, sd, min, and max
  meta-weight in the ensemble of each learner.
}
\description{
Warning: this function will eventually be moved into the
  SuperLearner package.
}

