% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_superlearner.R
\name{gen_superlearner}
\alias{gen_superlearner}
\title{Setup a SuperLearner() based on parallel configuration.}
\usage{
gen_superlearner(parallel = "multicore", cluster = NULL,
  outer_cv_folds = NULL, verbose = T)
}
\value{
A list with a SL and CV.SL function.
}
\description{
Setup a SuperLearner() based on parallel configuration.
}
