% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_java_memory.R
\name{get_java_memory}
\alias{get_java_memory}
\title{Return amount of RAM allocated to rJava}
\usage{
get_java_memory(verbose = FALSE)
}
\arguments{
\item{verbose}{If TRUE display more detailed information during execution.}
}
\description{
This presumes that the rJava package has already been loaded/initialized.
}
\examples{

get_java_memory()

get_java_memory(verbose = TRUE)

}
