% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipa_table.R
\name{ipa_table}
\alias{ipa_table}
\title{Generate a table of IPA scores by learner}
\usage{
ipa_table(x, ...)
}
\arguments{
\item{x}{CV.SuperLearner object}

\item{...}{Remaining arguments passed through to the methods.}
}
\description{
IPA = index of prediction accuracy (Kattan & Gerds 2018)
}
\details{
Supports CV.SuperLearner objects currently.
}
