% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all_code.R
\name{load_all_code}
\alias{load_all_code}
\title{Load all R files in the lib directory.}
\usage{
load_all_code(lib_dir = "lib", exclude_files = c("function_library.R"),
  file_pattern = "\\\\.R$", recursive = T, verbose = T)
}
\arguments{
\item{lib_dir}{Directory contains the source code files.}

\item{exclude_files}{Exclude a list of files; exclude function_library.R by
default because we presume that is the main R library file.}
}
\description{
Load all R files in the lib directory.
}
