% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_packages}
\alias{load_packages}
\title{Load a list of packages.}
\usage{
load_packages(pkgs = NULL, auto_install = F, update = F, verbose = F,
  ...)
}
\arguments{
\item{pkgs}{Character vector of packages to load.}

\item{auto_install}{Install any packages that could not be loaded.}

\item{update}{Update packages where possible.}

\item{verbose}{If T display more detailed information during execution.}

\item{...}{Any additional parameters to pass through to install.packages()}
}
\description{
Load packages and install them from CRAN if they aren't already available.
}
\examples{

# Load these 4 packages and install them if necessary.
load_packages(c("MASS", "SuperLearner", "tmle", "doParallel"), auto_install = TRUE)

}
