% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Compute the mode of a vector (can be multiple results).}
\usage{
Mode(x, exclude_na = TRUE)
}
\arguments{
\item{x}{vector}

\item{exclude_na}{Exclude NA from being a possible mode.}
}
\value{
Vector of modal values in arbitrary order.
}
\description{
Via http://stackoverflow.com/questions/2547402/is-there-a-built-in-function-for-finding-the-mode
}
\examples{

library(ck37r)

data(Boston, package = "MASS")

table(Boston$chas)

Mode(Boston$chas)

}
