% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot the ROC curve for an ensemble object.}
\usage{
plot_roc(x, ...)
}
\arguments{
\item{x}{Ensemble object.}

\item{...}{Remaining arguments passed through to the method.}
}
\description{
Supports CV.SuperLearner and SuperLearner objects.
}
