% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table.R
\name{plot_table}
\alias{plot_table}
\title{Plot a table of values using ggplot}
\usage{
plot_table(x, metric = "auc", sort = TRUE)
}
\arguments{
\item{x}{Table of ML results, assuming to have "learner", "ci_lower", and "ci_upper" columns.}

\item{metric}{Column to plot}

\item{sort}{If TRUE (default), sort rows based on the metric.}
}
\description{
Plot a table of values using ggplot
}
