% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-tree.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot rpart decision tree with pretty good defaults.}
\usage{
plot_tree(
  tree,
  shadow.col = "gray",
  shadow.offset = 0.2,
  box.palette = "auto",
  digits = 3,
  pal.thresh = 0.04,
  extra = 107,
  tweak = 1.2,
  ...
)
}
\arguments{
\item{tree}{tbd}

\item{shadow.col}{tbd}

\item{shadow.offset}{tbd}

\item{box.palette}{tbd}

\item{digits}{tbd}

\item{pal.thresh}{tbd}

\item{extra}{tbd}

\item{tweak}{tbd}

\item{...}{Additional arguments passed through to prp()}
}
\description{
Plot rpart decision tree with pretty good defaults.
}
