% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prauc_table.R
\name{prauc_table}
\alias{prauc_table}
\title{Generate a table of PR-AUCs by learner}
\usage{
prauc_table(x, ...)
}
\arguments{
\item{x}{SuperLearner or CV.SuperLearner object}

\item{...}{Remaining arguments passed through to the methods.}
}
\description{
Currently supports SuperLearner or CV.SuperLearner objects.
}
