% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_stratified.R
\name{predict.SL.stratified}
\alias{predict.SL.stratified}
\title{predict() for SL.stratified}
\usage{
\method{predict}{SL.stratified}(object, newdata, family, ...)
}
\arguments{
\item{object}{SuperLearner object}

\item{newdata}{Dataframe to predict the outcome}

\item{family}{"gaussian" for regression, "binomial" for binary
classification. (not useD)}

\item{...}{Additional arguments (not used)}
}
\description{
predict() for SL.stratified
}
