% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_count_terminal_nodes.R
\name{rf_count_terminal_nodes}
\alias{rf_count_terminal_nodes}
\title{Count the terminal nodes in each tree from a random forest}
\usage{
rf_count_terminal_nodes(rf)
}
\arguments{
\item{rf}{Random Forest object}
}
\value{
vector of terminal node counts
}
\description{
Count the terminal nodes in each tree from a random forest
}
