% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_table.R
\name{rmse_table}
\alias{rmse_table}
\title{Generate a table of RMSEs by learner}
\usage{
rmse_table(x, ...)
}
\arguments{
\item{x}{SuperLearner object}

\item{...}{Remaining arguments passed through to the methods.}
}
\description{
Supports SuperLearner objects currently.
}
