% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_auc.R
\name{se_auc}
\alias{se_auc}
\title{Standard error of the AUC}
\usage{
se_auc(auc, n_p, n_n)
}
\arguments{
\item{auc}{AUC estimate (numeric).}

\item{n_p}{Number of positive cases.}

\item{n_n}{Number of negative cases.}
}
\value{
Standard error of the AUC.
}
\description{
Computes the standard error of an AUC estimate using its equivalence to the
Wilcoxon statistic (Hanley & McNeil 1982). Adapted from
\code{auctestr::se_auc()} (MIT license, Copyright 2017 Josh Gardner), which
has been archived on CRAN.
}
\examples{
se_auc(0.75, 20, 200)
# Standard error decreases when classes are more balanced.
se_auc(0.75, 110, 110)
# Standard error increases when sample size shrinks.
se_auc(0.75, 20, 20)

}
\references{
Hanley and McNeil, The meaning and use of the area under a
  receiver operating characteristic (ROC) curve. Radiology (1982) 143 (1)
  pp. 29-36.
}
