% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_java_memory.R
\name{set_java_memory}
\alias{set_java_memory}
\title{Allocate memory to rJava}
\usage{
set_java_memory(memory = NULL, verbose = FALSE)
}
\arguments{
\item{memory}{Amount of memory to allocate to rJava, e.g. "4g".}

\item{verbose}{If TRUE display more detailed output during execution.}
}
\description{
This function should be run before rJava is loaded.
}
\examples{

# Set java maximum memory usage to 4 GB.
set_java_memory("4g", verbose = TRUE)

}
