% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_parallelism.R
\name{setup_parallelism}
\alias{setup_parallelism}
\title{Setup parallel processing, either multinode or multicore.}
\usage{
setup_parallelism(conf = NULL, type = "either", allow_multinode = T,
  machine_list = Sys.getenv("SLURM_NODELIST"),
  cpus_per_node = as.numeric(Sys.getenv("SLURM_CPUS_ON_NODE")),
  outfile = "")
}
\description{
By default it uses a multinode cluster if available, otherwise sets up multicore via doMC.
Libraries required: parallel, doSNOW, doMC, RhpcBLASctl, foreach
}

