% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_stratified.R
\name{SL.stratified}
\alias{SL.stratified}
\title{Stratified estimator}
\usage{
SL.stratified(Y, X, newX, family, obsWeights, id, stratify_on, ...)
}
\arguments{
\item{Y}{Outcome variable}

\item{X}{Covariate dataframe}

\item{newX}{Dataframe to predict the outcome}

\item{family}{"gaussian" for regression, "binomial" for binary
classification. Untested options: "multinomial" for multiple classification
or "mgaussian" for multiple response, "poisson" for non-negative outcome
with proportional mean and variance, "cox".}

\item{obsWeights}{Optional observation-level weights}

\item{id}{Optional id to group observations from the same unit (not used
currently).}

\item{stratify_on}{Vector of variables used to create stratification cells,
e.g. c('age', 'gender').}

\item{...}{Any other arguments, not used.}
}
\description{
This estimator stratifies the dataset on the values of the
specified variables and predicts the outcome mean within each cell. It is
intended as a low-variance, high bias estimator that can often provide
better predictions than the overall outcome mean.
}
