% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_auc.R
\name{sl_auc}
\alias{sl_auc}
\title{Calculate cross-validated AUC from CV.SuperLearner result}
\usage{
sl_auc(sl, Y = sl$Y)
}
\arguments{
\item{sl}{CV.SuperLearner object}

\item{Y}{Outcome vector, if not already added to SL object.}
}
\value{
List with cvAUC and ci elements.
}
\description{
Also calculates confidence interval and p-value.
Based on initial code by Alan Hubbard.
}
\references{
Add Erin LeDell chapter reference.
}
