% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_auc_table.R
\name{sl_auc_table}
\alias{sl_auc_table}
\title{Deprecated, use auc_table() instead.}
\usage{
sl_auc_table(sl, ...)
}
\arguments{
\item{sl}{SuperLearner object}

\item{...}{Remaining arguments that are passed through.}
}
\description{
Deprecated, use auc_table() instead.
}
