% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_plot_roc.R
\name{sl_plot_roc}
\alias{sl_plot_roc}
\title{Plot a ROC curve from cross-validated AUC from SuperLearner}
\usage{
sl_plot_roc(sl, Y = sl$Y, learner = which.min(sl$cvRisk),
  title = paste0("SuperLearner cross-validated ROC: ",
  names(sl$cvRisk)[learner]), digits = 4)
}
\arguments{
\item{sl}{SuperLearner object}

\item{Y}{tbd}

\item{learner}{Which learner to plot - defaults to minimum risk learner.}

\item{title}{tbd}

\item{digits}{tbd}
}
\description{
Thank you to Alan Hubbard for the initial code.
}
\references{
Add Erin LeDell paper/chapter.
}
