% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_stderr.R
\name{sl_stderr}
\alias{sl_stderr}
\title{Calculate the SE of individual SL learners}
\usage{
sl_stderr(sl, Y, obsWeights = rep(1, length(Y)))
}
\arguments{
\item{sl}{SuperLearner result object}

\item{Y}{Outcome vector}

\item{obsWeights}{Observation weights}
}
\description{
This will help understand risk estimates of learners in SL, similar to CV.SL.
}
