% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Rescale variables, possibly excluding some columns}
\usage{
standardize(x, ..., exclude_vars = NULL)
}
\arguments{
\item{x}{Dataframe or matrix, assumed to have column names.}

\item{...}{Extra arguments passed-through to base::scale}

\item{exclude_vars}{List of names of variables not to scale.}
}
\value{
Data-frame with appropriate variables scaled.
}
