% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_parallel.R
\name{tmle_parallel}
\alias{tmle_parallel}
\title{Modify TMLE to support parallel computation for g and Q.}
\usage{
tmle_parallel(Y, A, W, family, g.SL.library, Q.SL.library, id = 1:length(Y),
  verbose = F, gbound = 0.025, V = 5,
  sl_fn = SuperLearner::SuperLearner, ...)
}
\description{
This is needed to use Savio or any multicore system effectively.
}

