% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vim_corr.R
\name{vim_corr}
\alias{vim_corr}
\title{Correlation analysis}
\usage{
vim_corr(
  covariates,
  data,
  outcome,
  weights = rep(1, nrow(data)),
  bootse = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{covariates}{Character vector listing the covariates to analyze.}

\item{data}{Dataframe of covariates and outcome.}

\item{outcome}{Outcome variable, assumed binary.}

\item{weights}{Optional observation weight vector.}

\item{bootse}{Whether bootstrap standard errors should be used for inference.}

\item{verbose}{If TRUE display extra information.}
}
\description{
Examine variables most correlated with the outcome. Currently assumes that
outcome is binary variable.
}
