% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvsl_auc.R
\name{cvsl_auc}
\alias{cvsl_auc}
\title{Calculate cross-validated AUC from CV.SuperLearner result}
\usage{
cvsl_auc(cvsl)
}
\arguments{
\item{cvsl}{CV.SuperLearner object}
}
\value{
List with cvAUC and ci elements.
}
\description{
Also calculates confidence interval.
Based on initial code by Alan Hubbard.
}
\references{
Add Erin LeDell chapter reference.
}

