% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvsl_auc_table.R
\name{cvsl_auc_table}
\alias{cvsl_auc_table}
\title{Deprecated, use auc_table() instead.}
\usage{
cvsl_auc_table(cvsl, ...)
}
\arguments{
\item{cvsl}{CV.SuperLearner object}

\item{...}{Remaining arguments that are passed through.}
}
\description{
Deprecated, use auc_table() instead.
}
