% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvsl_plot_roc.R
\name{cvsl_plot_roc}
\alias{cvsl_plot_roc}
\title{Plot a ROC curve from cross-validated AUC from CV.SuperLearner}
\usage{
cvsl_plot_roc(cvsl, Y = cvsl$Y,
  title = "CV-SuperLearner Cross-validated ROC", digits = 4)
}
\arguments{
\item{cvsl}{tbd}

\item{Y}{tbd}

\item{title}{tbd}

\item{digits}{tbd}
}
\description{
Thank you to Alan Hubbard for the initial code.
}
\references{
LeDell, E., Petersen, M., & van der Laan, M. (2015). Computationally
efficient confidence intervals for cross-validated area under the ROC curve
estimates. Electronic journal of statistics, 9(1), 1583.

Polley EC, van der Laan MJ (2010) Super Learner in Prediction. U.C. Berkeley
Division of Biostatistics Working Paper Series. Paper 226.
http://biostats.bepress.com/ucbbiostat/paper266/

van der Laan, M. J., Polley, E. C. and Hubbard, A. E. (2007) Super Learner.
Statistical Applications of Genetics and Molecular Biology, 6, article 25.
http://www.degruyter.com/view/j/sagmb.2007.6.issue-1/sagmb.2007.6.1.1309/sagmb.2007.6.1.1309.xml
}
