% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvsl_plot_roc.R
\name{cvsl_plot_roc}
\alias{cvsl_plot_roc}
\title{Deprecated, please use plot_roc() instead.}
\usage{
cvsl_plot_roc(cvsl, ...)
}
\arguments{
\item{cvsl}{CV.SuperLearner object}

\item{...}{Remaining arguments that are passed through.}
}
\description{
Deprecated, please use plot_roc() instead.
}
