% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors_to_indicators.R
\name{factors_to_indicators}
\alias{factors_to_indicators}
\title{Convert factors to indicator variables.}
\usage{
factors_to_indicators(
  data,
  predictors = colnames(data),
  max_levels = 200L,
  verbose = FALSE
)
}
\arguments{
\item{data}{.}

\item{predictors}{.}

\item{max_levels}{If a factor contains more than this many levels, issue
a warning and don't convert it to indicators.}

\item{verbose}{TBD}
}
\description{
Turns factors into indicator variables with reasonable names.

Supports parallelization using \code{future} plans.

Based on code by Taylor Terry from 2013.
}
