% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_superlearner.R
\name{gen_superlearner}
\alias{gen_superlearner}
\title{Setup a SuperLearner() based on parallel configuration.}
\usage{
gen_superlearner(parallel = "multicore", cluster = NULL,
  outer_cv_folds = NULL, verbose = T)
}
\arguments{
\item{parallel}{Can be "multicore", "doSNOW"/"snow" or "seq"/NULL.}

\item{cluster}{Optional cluster, e.g. from SNOW.}

\item{outer_cv_folds}{How many folds to use for CV.SuperLearner}

\item{verbose}{If TRUE will output additional details during execution.}
}
\value{
A list with a SL and CV.SL function.
}
\description{
Create custom SL and CV.SL functions that automatically use
  parallelization based on the provided configuration.
}
