% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_values.R
\name{impute_missing_values}
\alias{impute_missing_values}
\title{Impute missing values in a dataframe and add missingness indicators.}
\usage{
impute_missing_values(data, add_indicators = T, prefix = "miss_",
  skip_vars = c())
}
\arguments{
\item{data}{Dataframe or matrix.}

\item{add_indicators}{Add a series of missingness indicators.}

\item{prefix}{String to add at the beginning of the name of each missingness
indicator.}

\item{skip_vars}{List of variable names to exclude from the imputation.}
}
\description{
Impute missing values in a dataframe and add missingness indicators.
}

