% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness_indicators.R
\name{missingness_indicators}
\alias{missingness_indicators}
\title{Return matrix of missingness indicators for a dataframe or matrix.}
\usage{
missingness_indicators(data, prefix = "miss_", remove_constant = T,
  remove_collinear = T, verbose = F)
}
\arguments{
\item{data}{Dataframe or matrix to analyze for missingness.}

\item{prefix}{Name prefix for new indicator columns.}

\item{remove_constant}{Remove any indicators that are all 0 or all 1.}

\item{remove_collinear}{Remove any indicators that are collinear with each other.}

\item{verbose}{If TRUE, print additional information.}
}
\value{
Matrix of missingness indicators, with any constant indicators removed.
}
\description{
Return matrix of missingness indicators for a dataframe or matrix.
}
