% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SuperLearner.R
\name{plot.SuperLearner}
\alias{plot.SuperLearner}
\title{Plot estimated risk and confidence interval for each learner}
\usage{
\method{plot}{SuperLearner}(sl, Y = sl$Y, constant = qnorm(0.975),
  sort = TRUE, ...)
}
\description{
Does not include SuperLearner or Discrete SL results as that requires
CV.SuperLearner to estimate the standard errors.
}
