% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prauc_table.SuperLearner.R
\name{prauc}
\alias{prauc}
\title{Calculate PR-AUC from cross-validation results}
\usage{
prauc(preds, actual, test_folds)
}
\arguments{
\item{preds}{Predicted values or score}

\item{actual}{Known label value}

\item{test_folds}{Length n vector assigning each vector to a given test fold}
}
\value{
Dataframe of results, including PR-AUC and standard deviation
}
\description{
Uses the precrec package to calculate PR-AUC
}
