% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Calculate RMSE from cross-validation results}
\usage{
rmse(preds, actual, test_folds)
}
\arguments{
\item{preds}{Predicted values or score}

\item{actual}{Known value}

\item{test_folds}{Length n vector assigning each vector to a given test fold}
}
\value{
Dataframe of results, including RMSE, standard error, and CI
}
\description{
Calculate RMSE from cross-validation results
}
\references{
Faber, N. K. M. (1999). Estimating the uncertainty in estimates of root mean
square error of prediction: application to determining the size of an
adequate test set in multivariate calibration. Chemometrics and Intelligent
Laboratory Systems, 49(1), 79-89.
}
