% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_auc.R
\name{sl_auc}
\alias{sl_auc}
\title{Table of cross-validated AUCs from SuperLearner result}
\usage{
sl_auc(sl, Y = sl$Y)
}
\arguments{
\item{sl}{CV.SuperLearner object}

\item{Y}{Outcome vector, if not already added to SL object.}
}
\value{
Dataframe table with auc, se, ci, and p-value (null hypothesis = 0.5).
}
\description{
Calculates cross-validated AUC for each learner in the SuperLearner. Also
calculates standard-error, confidence interval and p-value. Based on initial
code by Alan Hubbard.
}
\examples{
library(SuperLearner)
library(ck37r)

data(Boston, package = "MASS")

set.seed(1)
sl = SuperLearner(Boston$chas, subset(Boston, select = -chas), family = binomial(),
                 SL.library = c("SL.mean", "SL.glmnet"))

sl_auc(sl, Y = Boston$chas)

}
\references{
LeDell, E., Petersen, M., & van der Laan, M. (2015). Computationally
efficient confidence intervals for cross-validated area under the ROC curve
estimates. Electronic journal of statistics, 9(1), 1583.
}
