% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_plot_roc.R
\name{sl_plot_roc}
\alias{sl_plot_roc}
\title{Deprecated, please use plot_roc() instead.}
\usage{
sl_plot_roc(sl, ...)
}
\arguments{
\item{sl}{SuperLearner object}

\item{...}{Remaining arguments that are passed through.}
}
\description{
Deprecated, please use plot_roc() instead.
}
