% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_cluster.R
\name{stop_cluster}
\alias{stop_cluster}
\title{Stop the cluster if snow::makeCluster() was used, but nothing needed if doMC was used.}
\usage{
stop_cluster(cluster_obj)
}
\arguments{
\item{cluster_obj}{Cluster object generated by parallelize()}
}
\description{
Stop the cluster if snow::makeCluster() was used, but nothing needed if doMC was used.
}
