% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-vars.R
\name{summarize_vars}
\alias{summarize_vars}
\title{Summary table of variables}
\usage{
summarize_vars(
  df,
  vars = names(df),
  few_uniq_vals = 10L,
  groups = NULL,
  integers = NULL,
  ordinal = NULL
)
}
\arguments{
\item{df}{Dataframe of variables}

\item{vars}{Variable names to analyze in the dataframe}

\item{few_uniq_vals}{Number of values below which we display the numeric variable
as a candidate to be an integer.}

\item{groups}{TBD}

\item{integers}{Explicitly note integer-valued variables}

\item{ordinal}{Explicitly note ordinal variables}
}
\description{
Generate a summary table of variables in a dataframe.
}
